Arreglos a los test
- Eliminamos el getter innecesario
- Cambiamos los test de Lugar
- Creamos una clase por cada Test de Nodo
