+++ src/test/java/com/chucknorris/node/NodeTest.java
@@ -1,4 +1,4 @@
-package com.chucknorris.node.test;
+package com.chucknorris.node;
 
 import com.chucknorris.commons.Position;
 import com.chucknorris.gamemap.nodes.*;
@@ -12,12 +12,11 @@ import static org.junit.Assert.assertEquals;
 
 public class NodeTest {
 
-	Player p, p2;
+	Player p;
 
 	@Before
 	public void executedBeforeEach() {
 		p = new Player("Milei", 0);
-		p2 = new Player("Cristi", 100, 150);
 	}
 
 	@Test
@@ -66,21 +65,4 @@ public class NodeTest {
 		assertEquals("END", nodeTest.getType());
 	}
 
-	@Test
-	public void InvPesosPlusNodeTest() {
-		Node nodeTest = new InvPesosPlusNode(null, new Position(4, 4));
-
-		nodeTest.applyReward(p2, null, null);
-		assertEquals((int) p2.getPesos(), (int) 120);
-		assertEquals("IPP", nodeTest.getType());
-	}
-
-	@Test
-	public void InvPesosMinusTest() {
-		Node nodeTest = new InvPesosMinusNode(null, new Position(4, 4));
-
-		nodeTest.applyReward(p2, null, null);
-		assertEquals((int) p2.getPesos(), (int) 90);
-		assertEquals("IPM", nodeTest.getType());
-	}
 }
